+++ packages/core/src/api/exceptions/index.ts
@@ -2,4 +2,11 @@ export { defaultExceptionType } from './const';
 
 export { initializeExceptions } from './initialize';
 
-export type { ExceptionEvent, ExceptionStackFrame, ExceptionsAPI, PushExceptionOptions } from './types';
+export type {
+  ExceptionEvent,
+  ExceptionStackFrame,
+  ExceptionsAPI,
+  PushExceptionOptions,
+  ExtendedError,
+  Stacktrace,
+} from './types';