+++ packages/core/src/api/exceptions/types.ts
@@ -7,15 +7,22 @@ export interface ExceptionStackFrame {
   colno?: number;
   lineno?: number;
 }
+export interface ExtendedError extends Error {
+  columnNumber?: number;
+  framesToPop?: number;
+  stacktrace?: Error['stack'];
+}
+
+export interface Stacktrace {
+  frames: ExceptionStackFrame[];
+}
 
 export interface ExceptionEvent {
   timestamp: string;
   type: string;
   value: string;
 
-  stacktrace?: {
-    frames: ExceptionStackFrame[];
-  };
+  stacktrace?: Stacktrace;
   trace?: TraceContext;
 }
 
@@ -25,5 +32,6 @@ export interface PushExceptionOptions {
 }
 
 export interface ExceptionsAPI {
+  pushError: (error: Error) => void;
   pushException: (value: string, options?: PushExceptionOptions) => void;
 }