+++ packages/core/src/api/initialize.ts
@@ -1,3 +1,4 @@
+import type { Config } from '../config';
 import type { Metas } from '../metas';
 import type { Transports } from '../transports';
 import { initializeExceptions } from './exceptions';
@@ -7,12 +8,12 @@ import { initializeMeta } from './meta/initialize';
 import { initializeTraces } from './traces';
 import type { API } from './types';
 
-export function initializeAPI(transports: Transports, metas: Metas): API {
+export function initializeAPI(config: Config, transports: Transports, metas: Metas): API {
   const tracesApi = initializeTraces(transports, metas);
 
   return {
     ...tracesApi,
-    ...initializeExceptions(transports, metas, tracesApi),
+    ...initializeExceptions(config, transports, metas, tracesApi),
     ...initializeMeta(transports, metas),
     ...initializeLogs(transports, metas, tracesApi),
     ...initializeMeasurements(transports, metas, tracesApi),