+++ packages/core/src/config/types.ts
@@ -1,4 +1,4 @@
-import type { APIEvent } from '../api';
+import type { APIEvent, ExtendedError, Stacktrace } from '../api';
 import type { Instrumentation } from '../instrumentations';
 import type { App, User, Session, MetaItem } from '../metas';
 import type { BeforeSendHook, Transport } from '../transports';
@@ -14,6 +14,7 @@ export interface Config<P = APIEvent> {
   user?: User;
   beforeSend?: BeforeSendHook<P>;
   ignoreErrors?: Patterns;
+  parseStacktrace?: (err: ExtendedError) => Stacktrace;
 }
 
 export type Patterns = Array<string | RegExp>;