+++ packages/web/src/instrumentations/errors/index.ts
@@ -3,6 +3,8 @@ import { BaseInstrumentation, VERSION } from '@grafana/agent-core';
 import { registerOnerror } from './registerOnerror';
 import { registerOnunhandledrejection } from './registerOnunhandledrejection';
 
+export { parseStacktrace } from './stackFrames';
+
 export class ErrorsInstrumentation extends BaseInstrumentation {
   readonly version = VERSION;
   readonly name = '@grafana/agent-web:instrumentation-errors';